cleanup docker registry
