+++ internal/registry/docker/docker.go
@@ -1,7 +1,6 @@
 package docker
 
 import (
-	"context"
 	"crypto/sha256"
 	"encoding/json"
 	"fmt"
@@ -15,7 +14,6 @@ import (
 	"github.com/gorilla/mux"
 
 	"github.com/valist-io/valist/internal/core/types"
-	"github.com/valist-io/valist/internal/storage"
 )
 
 type handler struct {
@@ -44,76 +42,6 @@ func NewHandler(client types.CoreAPI) http.Handler {
 	// POST /v2/<name>/blobs/uploads/?mount=<digest>&from=<repository name>
 
 	return handlers.LoggingHandler(os.Stdout, router)
-}
-
-func (h *handler) writeUpload(uuid string, r io.Reader) (int64, error) {
-	path := filepath.Join(os.TempDir(), uuid, "blob")
-
-	blob, err := os.OpenFile(path, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		return 0, err
-	}
-	defer blob.Close()
-
-	size, err := io.Copy(blob, r)
-	if err != nil {
-		return 0, err
-	}
-
-	h.uploads[uuid] += size
-	return size, nil
-}
-
-func (h *handler) findBlob(ctx context.Context, orgName, repoName, digest string) (string, error) {
-	if p, ok := h.blobs[digest]; ok {
-		return p, nil
-	}
-
-	raw := fmt.Sprintf("%s/%s/latest", orgName, repoName)
-	res, err := h.client.ResolvePath(ctx, raw)
-	if err != nil {
-		return "", err
-	}
-
-	return fmt.Sprintf("%s/%s", res.Release.ReleaseCID, digest), nil
-}
-
-func (h *handler) loadBlob(ctx context.Context, orgName, repoName, digest string) (storage.File, error) {
-	p, err := h.findBlob(ctx, orgName, repoName, digest)
-	if err != nil {
-		return nil, err
-	}
-
-	return h.client.Storage().Open(ctx, p)
-}
-
-func (h *handler) moveBlob(ctx context.Context, dir storage.Directory, orgName, repoName, digest string) error {
-	p, err := h.findBlob(ctx, orgName, repoName, digest)
-	if err != nil {
-		return err
-	}
-
-	if err := dir.Add(ctx, digest, p); err != nil {
-		return err
-	}
-
-	delete(h.blobs, digest)
-	return nil
-}
-
-func (h *handler) loadManifest(ctx context.Context, orgName, repoName, ref string) (storage.File, error) {
-	raw := fmt.Sprintf("%s/%s/latest", orgName, repoName)
-	res, err := h.client.ResolvePath(ctx, raw)
-	if err != nil {
-		return nil, err
-	}
-
-	release, err := h.client.GetRelease(ctx, res.Organization.ID, res.Repository.Name, ref)
-	if err == nil {
-		return h.client.Storage().Open(ctx, release.MetaCID)
-	}
-
-	return h.client.Storage().Open(ctx, fmt.Sprintf("%s/%s", res.Release.ReleaseCID, ref))
 }
 
 func (h *handler) getVersion(w http.ResponseWriter, req *http.Request) {
@@ -180,18 +108,15 @@ func (h *handler) patchBlob(w http.ResponseWriter, req *http.Request) {
 	orgName := vars["org"]
 	repoName := vars["repo"]
 
-	// TODO verify Range header is valid and return 416 if not
 	start := h.uploads[uuid]
-
-	size, err := h.writeUpload(uuid, req.Body)
-	if err != nil {
+	if err := h.writeBlob(uuid, req.Body); err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
 
-	w.Header().Set("Location", fmt.Sprintf("/v2/%s/%s/blobs/uploads/%s", orgName, repoName, uuid))
-	w.Header().Set("Range", fmt.Sprintf("%d-%d", start, start+size))
 	w.Header().Set("Docker-Upload-UUID", uuid)
+	w.Header().Set("Location", fmt.Sprintf("/v2/%s/%s/blobs/uploads/%s", orgName, repoName, uuid))
+	w.Header().Set("Range", fmt.Sprintf("%d-%d", start, h.uploads[uuid]))
 	w.WriteHeader(http.StatusAccepted)
 }
 
@@ -203,12 +128,11 @@ func (h *handler) putBlob(w http.ResponseWriter, req *http.Request) {
 	orgName := vars["org"]
 	repoName := vars["repo"]
 
-	// TODO verify Range header is valid and return 416 if not
+	// TODO calculate and compare digest
 	digest := req.URL.Query().Get("digest")
 	path := filepath.Join(os.TempDir(), uuid, "blob")
 
-	_, err := h.writeUpload(uuid, req.Body)
-	if err != nil {
+	if err := h.writeBlob(uuid, req.Body); err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
@@ -223,8 +147,8 @@ func (h *handler) putBlob(w http.ResponseWriter, req *http.Request) {
 	h.blobs[digest] = p
 	delete(h.uploads, uuid)
 
-	w.Header().Set("Location", fmt.Sprintf("/v2/%s/%s/blobs/uploads/%s", orgName, repoName, uuid))
 	w.Header().Set("Docker-Content-Digest", digest)
+	w.Header().Set("Location", fmt.Sprintf("/v2/%s/%s/blobs/uploads/%s", orgName, repoName, uuid))
 	w.WriteHeader(http.StatusAccepted)
 }
 
@@ -269,23 +193,26 @@ func (h *handler) putManifest(w http.ResponseWriter, req *http.Request) {
 		return
 	}
 
-	err = h.moveBlob(ctx, dir, orgName, repoName, manifest.Config.Digest)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
+	// add manifest to release directory
 	if err := dir.Add(ctx, digest, manifestCID); err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
 
-	for _, layer := range manifest.Layers {
-		err := h.moveBlob(ctx, dir, orgName, repoName, layer.Digest)
+	// add layers and config to release directory
+	for _, digest := range manifest.Digests() {
+		p, err := h.findBlob(ctx, orgName, repoName, digest)
 		if err != nil {
 			http.Error(w, err.Error(), http.StatusBadRequest)
 			return
 		}
+
+		if err := dir.Add(ctx, digest, p); err != nil {
+			http.Error(w, err.Error(), http.StatusBadRequest)
+			return
+		}
+
+		delete(h.blobs, digest)
 	}
 
 	release := &types.Release{
